projeto_inicial_drf_clientes-master/requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
Django==3.0.8
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.11.0
Faker==9.5.1
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.9.1
pylint==2.5.3
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
validate-docbr==1.9.0
wrapt==1.12.1
